refactor: one definition of "serialise, or refuse if it does not fit"
The measure-then-bound-then-serialise helper existed three times: byte-for-byte
in outputs/json_util.h and config/configuration.cpp, and open-coded a third time
in serializeConfigForStorage. All three enforce the same rule -- a document that
exceeds its ceiling is refused, never truncated, because a truncated JSON blob is
not a smaller answer but an unparseable one.

Moved to src/json_limits.h. At the root rather than under outputs/ or config/
because both need it and neither owns it -- which is exactly why it was
duplicated: there was no such place, and config/ writing the NVS blob could not
reach a helper filed under outputs/. It depends on ArduinoJson and nothing else,
which is what makes it placeable there at all.

json_util.h re-exports it so the twenty-odd payload builders keep reading the
same. serializeConfigForStorage now calls it instead of open-coding the checks,
which also gives it the doc.overflowed() and untouched-on-failure guarantees the
other two already had.

No behaviour change: 730 host tests pass unchanged, all four firmware envs build.
